Reorganize code into library and binary
