+++ Cargo.toml
@@ -3,13 +3,15 @@ name = "livetagger"
 version = "0.1.0"
 edition = "2024"
 
+[workspace]
+resolver = "3"
+members = ["livemeta"]
+
 [dependencies]
+livemeta = { path = "livemeta" }
+
 anyhow = "1.0.98"
 clap = { version = "4.5.37", features = ["derive"] }
 glob = "0.3.2"
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
-xmp_toolkit = "1.9.2"
-
-[dev-dependencies]
-pretty_assertions = "1.4.1"
