+++ livemeta/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "livemeta"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+xmp_toolkit = "1.9.2"
+
+[dev-dependencies]
+pretty_assertions = "1.4.1"
