+++ livemeta/src/test_data/tags_removed.xml
