+++ livemeta/src/test_data/tags_removed_all.xml
